Files Uploaded by AbuDarda
algorithm1 : Algorithm to take an array input from a text file and then sort it. A membership Query is performed and the output is written to another text file.

testgen_random : produces random values for for the array elements and the key. These values are written to text files.

oracle_random_test : reads values from the output text file algorithm1 and compares it with the outputs from a reference algorithm (linear search) and thus finds the number of failures of the test algorithm as  compared to reference algorithm.